Move api tests to /tests folder
