+++ app/lib/osm.js
@@ -7,7 +7,7 @@ const passport = require('passport-light')
 const R = require('ramda')
 const hydra = require('./hydra')
 const url = require('url')
-const db = require('../db')
+const db = require('../../lib/db')
 const xml2js = require('xml2js')
 const InternalOAuthError = require('passport-oauth').InternalOAuthError
 const OSMStrategy = require('passport-openstreetmap').Strategy
