+++ app/manage/permissions/clients.js
@@ -1,4 +1,4 @@
-const db = require('../../db')
+const db = require('../../../lib/db')
 
 /**
  * clients
