+++ app/manage/permissions/delete-client.js
@@ -1,4 +1,4 @@
-const db = require('../../db')
+const db = require('../../../lib/db')
 
 /**
  * client:delete
