+++ app/manage/permissions/index.js
@@ -1,4 +1,4 @@
-const db = require('../../db')
+const db = require('../../../lib/db')
 const hydra = require('../../lib/hydra')
 const { mergeAll, isNil } = require('ramda')
