--- app/tests/README.md
@@ -1,28 +1,0 @@
-# Tests
-
-Tests are run using [ava](https://npmjs.com/ava).
-
-Linting is performed using [standard](https://npmjs.com/standard)
-
-## Running tests
-
-```sh
-git clone {this repo}
-cd {this repo}
-npm install
-npm test
-```
-
-`npm test` runs only the tests.
-
-### Run the linter
-
-```sh
-npm run lint
-```
-
-### Run both the tests and the linter
-
-```sh
-npm run test-lint
-```
