+++ lib/db.js
@@ -1,5 +1,5 @@
 const knex = require('knex')
-const connections = require('./../../knexfile')
+const connections = require('../knexfile')
 
 const config = connections[process.env.NODE_ENV]
