+++ package.json
@@ -21,7 +21,7 @@
     "docs": "npm run docs:validate && npm run docs:api",
     "dev": "NODE_ENV=development nodemon --watch app app/index.js",
     "migrate": "knex migrate:latest",
-    "test": "NODE_ENV=test nyc ava app/tests/**/*.test.js -c 1 --serial --verbose",
+    "test": "NODE_ENV=test nyc ava tests/**/*.test.js -c 1 --serial --verbose",
     "lint": "next lint",
     "build": "next build",
     "start": "NODE_ENV=production node app/index.js",
